hello

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-unsigned long int a, b, c, d;
+long double a, b, c, d;
 b = 1;
 c = 0;
 for (a = 0; a < 98; a++)
@@ -14,7 +14,7 @@ for (a = 0; a < 98; a++)
 d = c + b;
 c = b;
 b = d;
-printf("%lu", d);
+printf("%0.0Lf", d);
 if (a != 97)
 printf(", ");
 }
